# frontend_2_0/public/code-docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "МосТранспорт — API диспетчерской",
    "description": "\nБэкенд диспетчерской системы прогноза задержек городского транспорта.\n\n**Поток данных.** Приёмник NDTP (TCP, порт 9201) разбирает навигационные пакеты терминалов и хранит\nисторию каждого ТС за последний час. Раз в 30 с цикл прогноза отправляет историю и плановое расписание\nв ML-сервис и получает прогноз задержки на первой остановке, плановое прибытие на которую через\n**10–15 минут**. Состояние ТС вместе с прогнозами рассылается дашборду по WebSocket раз в секунду.\n\n**С чего начать:**\n* `GET /api/telemetry/stats` — идёт ли поток NDTP;\n* `GET /api/predictions` — текущие прогнозы и состояние цикла прогноза (`elapsed_s`, ошибки ML);\n* `GET /api/predictions/{tr_id}/history` — ряд прогнозов одного ТС за час.\n\n**Уровень риска** (`risk`) по прогнозу задержки на целевой остановке:\n`late` — больше +3 мин, `risk` — от +1 до +3 мин, `on_time` — в пределах ±1 мин, `early` — раньше графика больше чем на 1 мин.\n\nВсе времена — ISO 8601 в UTC (`Z`); дашборд показывает их по Москве.\n\n**WebSocket** `/ws/vehicles` (в Swagger не отображается): при подключении и далее раз в секунду\nприходит JSON как у `GET /api/telemetry`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Вход диспетчера",
        "description": "``true``, если логин и пароль совпали с учётной записью в базе (bcrypt), иначе ``false``.",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Login Api Login Post"
                },
                "example": true
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/predictions/{tr_id}/history": {
      "get": {
        "tags": [
          "Прогноз"
        ],
        "summary": "Ряд прогнозов ТС за час",
        "description": "Прогнозы задержки ТС за последний час по одному на момент T — данные графика отклонения.",
        "operationId": "prediction_history_api_predictions__tr_id__history_get",
        "parameters": [
          {
            "name": "tr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tr Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Prediction History Api Predictions  Tr Id  History Get"
                },
                "example": [
                  {
                    "at": "2026-01-06T05:48:05Z",
                    "delay_s": 22,
                    "risk": "on_time"
                  },
                  {
                    "at": "2026-01-06T05:53:19Z",
                    "delay_s": 76,
                    "risk": "risk"
                  }
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/emulator/start": {
      "post": {
        "tags": [
          "Эмулятор"
        ],
        "summary": "Запустить эмулятор NDTP",
        "description": "Настраивает эмулятор организаторов слать пакеты сгенерированных терминалов на приёмник этого бэкенда.",
        "operationId": "emulator_start_api_emulator_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmulatorStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Emulator Start Api Emulator Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EmulatorStart": {
        "properties": {
          "units": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Units",
            "description": "Сколько терминалов сгенерировать",
            "default": 16
          },
          "interval_ms": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Interval Ms",
            "description": "Период отправки пакетов каждым терминалом, мс",
            "default": 1000
          }
        },
        "type": "object",
        "title": "EmulatorStart",
        "description": "Параметры запуска эмулятора NDTP."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginData": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login",
            "description": "Логин диспетчера"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Пароль"
          }
        },
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "title": "LoginData",
        "description": "Учётные данные диспетчера."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}